[Bronze II] Title: 숫자, Time: 832 ms, Memory: 27724 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/10093. 숫자/숫자.java
@@ -6,17 +6,17 @@ public class Main {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         String[] input = br.readLine().split(" ");
-        int a = Integer.parseInt(input[0]);
-        int b = Integer.parseInt(input[1]);
+        long a = Long.parseLong(input[0]);
+        long b = Long.parseLong(input[1]);
 
         if (a > b) {
             System.out.println(a-b-1);
-            for (int i = b + 1; i < a; i++) {
+            for (long i = b + 1; i < a; i++) {
                 System.out.print(i + " ");
             }
         } else if (a < b) {
             System.out.println(b-a-1);
-            for (int i = a + 1; i < b; i++) {
+            for (long i = a + 1; i < b; i++) {
                 System.out.print(i + " ");
             }
         } else {
